Remove unused web shim
